🌍 #138 Rename Gen.Create<T> => Gen.Auto<T>
Also gen factories into different files.
