+++ tests/GalaxyCheck.Tests.V2/GenTests/AutoGenTests/AboutGeneratingPrimitives.cs
@@ -10,7 +10,7 @@ using Xunit;
 using Property = NebulaCheck.Property;
 using Test = NebulaCheck.Test;
 
-namespace Tests.V2.GenTests.CreateGenTests
+namespace Tests.V2.GenTests.AutoGenTests
 {
     public class AboutGeneratingPrimitives
     {
@@ -23,7 +23,7 @@ namespace Tests.V2.GenTests.CreateGenTests
                 List<int> SampleTraversal(GalaxyCheck.IGen<int> gen) =>
                     AboutGeneratingPrimitives.SampleTraversal(gen, seed, size);
 
-                var gen0 = GalaxyCheck.Gen.Create<int>();
+                var gen0 = GalaxyCheck.Gen.Auto<int>();
                 var gen1 = GalaxyCheck.Gen.Int32();
 
                 var sample0 = SampleTraversal(gen0);
@@ -41,7 +41,7 @@ namespace Tests.V2.GenTests.CreateGenTests
                 List<char> SampleTraversal(GalaxyCheck.IGen<char> gen) =>
                     AboutGeneratingPrimitives.SampleTraversal(gen, seed, size);
 
-                var gen0 = GalaxyCheck.Gen.Create<char>();
+                var gen0 = GalaxyCheck.Gen.Auto<char>();
                 var gen1 = GalaxyCheck.Gen.Char();
 
                 var sample0 = SampleTraversal(gen0);
@@ -59,7 +59,7 @@ namespace Tests.V2.GenTests.CreateGenTests
                 List<string> SampleTraversal(GalaxyCheck.IGen<string> gen) =>
                     AboutGeneratingPrimitives.SampleTraversal(gen, seed, size);
 
-                var gen0 = GalaxyCheck.Gen.Create<string>();
+                var gen0 = GalaxyCheck.Gen.Auto<string>();
                 var gen1 = GalaxyCheck.Gen.String();
 
                 var sample0 = SampleTraversal(gen0);
